fix a bug with accidental form submit
This one was quite tricky to debug. I'm not 100% sure, but I think this
bug was caused by the clickable spiciness labels which were button elements.

Button elements seems to default to "type=submit" by default when they are
a form's children. So explicitly setting them to "type=button" seems to
fix this issue.